Initial checkin - build VS 2017 solution.
After building the project once in a completely different checkout and
figuring out the ins and outs, this is a second stab to get it cleanly.
This is going to attempt to use vcpkg to have an easier to build project.

Note: this will be Windows specific and over time, I will delete other
platforms if they clutter global searches / they just aren't used to make
a concise repository to work in.

Next steps:
* install dependencies using a nupkg generated by vcpkg
* update to build for Win 10 Anniversary update (14393)
* fix build issues