Update library names and paths
This is temporary. These libraries come from the beta of vcpkg. It made things much eaiser to build but there are bugs. It should have made it possible to not have to modify some of this (the library names were a must) but the paths not according to my understanding of the tool.

In addition, there were some packages that i had to build against the debug version of the lib whereas the rest build against the release versions. That seems wrong. Note: I'm referring here to having to load the "d" versions of dlls at runtime.

Bugs filed for all of these:
Bug 32
Bug 31
Bug 30